Fixed assertions from GMime.
I'm not 100% sure if this is correct behaviour, but testing
for GMIME_PART on the part-object before using it does resolve
the scary-assertions.

This code is relating to whether we show inline-attachments
via the lua boolean variable:

        view_inline_attachments

(Defaults to true, even though I think it is icky.  This is one
of those changes I wish I'd not agreed to.)

This closes #223.