Initial stub at getting MIME-parts accessible to Lua.
This seems to work well, but as per the noted TODOs there are
some omissions.  Most obvious is the parts() binding to Lua,
that returns a nested table.

Instead message:parts() should return an array of CMessagePart
objects, and those objects should be wrapped suitably to Lua.

To test this code please run:

        $ make
        $ ./lumail parts.lua

This will use the pre-cooked Maildir/ entries which are now
stored within this repository.