Added virtual view-classes for UI display.
This is still work in-progress, but we've setup a base-class
which suffices to allow distinct-views to be instantiated and
used to display different things.

This updates #7.
This updates #2.

The next step is to expand the display of the screen display
with proper handling, and a decent way of entering lua code to
be executed.

This code will be built around that in `misc/ui-demo.c`.