Display a message - via Lua callback.
This commit actually makes the client somewhat convincing:

* When in message-mode draw a message on the screen.
* Draw the message by invoking the user-define `CMessage:to_string` function.

I'd suggest instead that the `to_string` method should return
an array/table, instead of just a single string.  But that can
come in the future.

I can imagine seeing the `to_string` method returning something like:

     { "${YELLOW}To: steve@steve.org.uk",
       "${RED}From: bob@example.com", ...
       ...
       "Bob",
       "-- ",
     }

The downside is that if we let the user write that it might get
complicated.  I'm not sure.

Closes #2.