We have a working mail-viewing application.
This mamoth set of commits is not to be repeated, but because it
involved so many coordinated changes it was a useful way to
spend the night.

The short version is that we now have a few more primitives
and our global-state object can get/set the selected maildir,
and message.

The `dump_maildir.lua` sample-code shows how things work,
roughly.

The next set of changes will be more simple, but we probably
need to add CMaildir::get_messages() to lua, to allow simpler
iteration.

We're also missing things like CMessage:delete().

This closes #10.