The drawing code for all our modes is in one place.
This update changes all the view-modes to use one routine to
draw the output of that mode.

The observation behind this is simple:

* Every mode draws an array of lines.
* Those lines might be CMaildirs, CMessages, or message-bodies.
* Regardless the ultimate aim is to fetch the stuff, then draw it, line by line.

With that in mind each mode has been updated to fetch an array of
"stuff to draw", then the CScreen mode has a single, identical, simple
method to draw that array of "stuff".

This closes #21.