Update the drawing of messages.
In the past we'd retrieve the current message and call the Lua
`Message.to_string()` function against it.  This was a mistake,
instead now we just call:

        message_view()

This `message_view()` function, implemented in Lua, should find the
current message AND return the content of it, suitable for formatting.

This updates #30.