Updated CConfig class.
This update switches to using a `map` for storage, rather than
a vector.  This increases speed and is more sensible anyway.

We've also added the ability for a CConfigEntry to store an
integer, in addition to a string, or array of strings.

This closes #56, but I will not push since this change breaks
the client.  (i.e. Other things need to change for this to be
a useful commit.)