Properly draw multiple colours.
This updates closes #61, and is demonstrated in `lua_view()`.  To view
the new code in action run:

                ./lumail2 --load-file ./lumail2.lua

Then press `L` and look at the first line.

We corectly switch colours mid-line, add padding, and cope with
horizontal scrolling.

The only outstanding issue is to do the same to the panel-code.