Use our new Lua-helper function.
We now use the single method which allows us to get a table
from a Lua function.  This will reduce duplication and the
chance for errors.

This is particularly good because we think we have a bug
in the preivous version - we didn't handle the error value
that `lua_pcall` might have returned, which could lead
to the segfault we see in #78.