Avoid potentially invalid memory.
@jugglerchris points out that after popping a string from
the stack it will no longer be valid.

We have:

                *string

We want:

                *function
                *string

But we can't just call `get_global` to find the function,
and then swap the arguments because the get_global function
might fail - if the usr doesn't define `on_error`.

So explicitly copy the string, then free it post-use.