Avoid screen flicker.
Every time around the screen loop we used to run:

* clear the screen
* refresh
* draw the lines.
* refresh.

This caused flickering, which was most obvious in the (re)display
of the panel - because the box-characters would flash.

Updated the `clear()` method called in step one to not refresh
meaning that the screen clearing still takes effect, but doesn't
trigger an immediate refresh.

This closes #89.