Updated to handle message-parsing failures.
This update corrects a bug in the previous revision, where we
closed the file-descriptor relating to the message, causing
all part-reads to fail.

We now skip multiple/alternative parts, as being things that
can't have content - only children - and this means that we
can correctly read, process, and display the sample maildir
supplied.  (Both correctly and without a segfault.)

This closes #98.