Correctly parse messages to a tree of CMessageParts.
This update closes #101, by allowing a message to be parsed into
multiple MIME-parts which are stored as a tree.   (Specifically a
MIME-part can have children; this is a parent -> child relationship
the child has no access to the parent.)

There is probably work to be done to improve things, but for the
moment test by viewing a message and pressing "`A`" to see
attachment-mode - which will have the lovely tree-structure in place.