Added Screen:stuff(txt).
This allows input to be placed into the buffer from which all
input is read - allowing automation of user-actions.

I'm not sure how useful it is, but it was something that we supported
in the past, and it seems like it might be of use to some users.

This updates and closes #104.