Wire up parent-relationship to Lua.
This doesn't quite work the way that the sample code should
have done - because two user-data objects representing the same
thing have different IDs.

This is best-demonstrated like so:

      x1 = Message.new( "/home/steve/Maildir/attachment/cur/1441627138.23705_2.ssh.steve.org.uk:2,S")
      x2 = Message.new( "/home/steve/Maildir/attachment/cur/1441627138.23705_2.ssh.steve.org.uk:2,S")
      print( x1, x2 )
      assert(x1 == x2 )

      userdata: 0x199ab98userdata: 0x199dcc8
      ERROR ./tt.lua:4: assertion failed!

Same failure-case as here, even though logically we've got two
references to the same thing the pointers are different so comparison
fails.

This closes #111.