Fixed horizontal-scrolling.
This commit changes the way that lines of text are tokenized, such
that we have a lot more repetition.

In the past we would split the string "$[RED]This is red" into a single
entry of COLOUR_STRING containing:

        colour => RED
        text -> "This is red"

What this commit does is allow us to update that tokenized entry
and *repeat* it, once for each column in the display.  This means
we'll have entries like so:

        colour = RED, text => T
        colour = RED, text => h
        colour = RED, text => i
        colour = RED, text => s
        colour = RED, text =>
        colour = RED, text => i
        colour = RED, text => s
        colour = RED, text =>
        colour = RED, text => r
        colour = RED, text => e
        colour = RED, text => d

This repetition means that we can then "scroll" by starting from
the scroll-offset in our list, and we *don't* need to keep track
of where we are in the text, or perform complex calculations:

* This update is provably correct, and naive.
* Naive is simple.
* Simple is good.