Rework the way messages are sent.
Rather than saving a message being a property/action carried out
against a message it is now an operation carried out against
a maildir.

This means we now write code like so:

        m = Message.new( "path/to/file" )
        f = Maildir.new( "/home/steve/Maildir/foo/" )
        f.save_message(m);

This might seem like an arbitrary distinction but it means
that folders know how to save messages - and messages don't
need to care.  This will mean CMessage won't need to be updated
when we allow IMAP saving ..

This closes #118.