Adjust unread-message count manually.
Since the count of total/unread messages is only fetched for IMAP
messages when the CMaildir object is created - if we change state
in the same run we don't update things.

This is a problem which might require a rethink of how IMAP operations
work - and the interaction between CMaildirs and CMessages.

(The ideal case CMaildir would get the messages from "itself", via
IMAP.  That can't happen because CMaildirs are created/destroyed
fairly frequently via CGlobalState.)

This closes #135.