Updated handling of colour-strings.
We've now updated things such that we can parse $[#RED] as
well as $[RED] (i.e. added "#" to the regexp for valid ids).

WHen we see a colour-string which is prefixed with the # we
use the _previous_ colour, defaulting to white if this is the
first part of the line, and rework the text such that we display
the unescaped version.

This updates #121.