Correctly bound line-drawing to avoid wrapping.
This update ensures that we never draw a line of text
wider than the width of the screen, which can cause
ugly output.

We take advantage of the fact that we split our
coloured-strings into one-character segments to skip
drawing them once we've exceeded the width.

This does mean that we do more work though, I think the
amount of work is minimal.

We convert, for example:

   [$RED]This is red $[BLUE]Blue!

Into

    T
    h
    i
    s
    [space]
    i
    s
    [space]
    r
    e
    d
    [space]
    B
    l
    u
    e

we avoided this in the past, unless horizontal scrolling was
in effect, but now we do it unconditionally because this allows
us to easily work out the width of the string.

NOTE: We have to disable this behaviour if line-wraping is
enabled, which is a minor complication.

This closes #156.