luamail2.lua: fix get_mime_viewer()
io.lines() is called with the argument "file" which is not defined.

io.input() sets the opened file as default file descriptor, so io.lines()
needs no argument at all.

Also store the path to the mailcap file in the local variable "mailcap_file"
for easier maintenance and modification by the user.