Rethink message-saving.
I've tied myself in knots thinking about this, but the short
version is that the simplest solution to the problem is to
make saving messages save them as "read".

Because the most obvious case that is the correct thing to do,
and in the atypical case it can be corrected by the user easily
enough - although it is true the API could make that simpler it
isn't required.

This closes #160.