Revert the attempted optimisation.
We attempted to avoid updating our cache if the same thing
was selected again, however this caused the crash seen in

It could have been possible to fix this, by testing the new
thing for being non-NULL, however this would have resulted
in bogus results -as the bug report shows it is triggered
when a deletion happens, and if we don't update our cache
when a message is deleted then we'll have a phantom
message left hanging around.

So the correct solution is to not cache, or have a better
cache invalidation strategy...

This closes #174.