Reordered our maildir/message operations.
This is part of the performance scan in #173, and is based
upon the observation of how things are executed:

* First we get all the messages|maildirs
* Then we sort them.
* Then we apply the limits.

If we limit _then_ sort we have to sort fewer items, which
is a bigger optimization in the general case.  (Although
the case of limit=="today" we probably don't gain as much
the other cases we do.)

This updates #173.