Default to selecting the max message.
This is a matter of choice, but it tends to be the most
sane choice - we'd expect to highlight new messages,
which are the most recent, so with the default sorting
set to `file` we'd have those at the bottom.

This is an arbitrary choice:

* Open folder; select first message.
vs.
* Open folder; select last message.

I will not make this user-selectable via a flag though because
this used to be the default and too many config settings are
a mistake.

Closes #175.