Implement Global:sort_messages().
This is the final part of our sorting change, it adds a new
method to the Lua API:

    Global:sort_messages()

This method takes a table of CMessage objects, and returns
that table in a sorted order.  The sorting is achieved by
calling a Lua function `compare_by_XX`, where `XX` is retrieved
by looking up the value of the `index.sort` method.

This closes #176.