Added mimeparts2table function.
This function takes a message and returns a table containing
each of the MIME-parts.  THis table is then the sole function
for parsing MIME-parts, as used in the attachment-mode.

By centralizing this routine we've improved things significantly
by cutting down duplication, and as a bonus this function contains
a nested function which recursively calls itself, which means that
it handles all types of nested and nested-nested MIME parts.

This updates #181.

NOTE: It might be this implementation will be improved, to show
the level of nesting, via @jugglerchris.

NOTE: We still need to update `message_view()` to call this
function - to display text in the message-view.