Updated how we fetch message-body for display.
This uses the new mimeparts2table routine to parse the message
into MIME parts, and builds up the possible alternatives via
that mechanism:

* Keeps a running track of text/plain and text/html parts.
* Appends/Prepends parts if appropriate.

If both are empty we're done.

Otherwise pick the first that makes sense.

This updates #181, and probably resolves it, but I'll leave the
issue open for the moment just in case Chris untangles his mime
parsing code - and I can use that to update/replace the mimepart2table
routine this commit and the previous one relies upon.