Allow lua to draw to the screen.
This is naive because if you draw inside a view-mode chances
are your text will not be shown.

In short it relies upon the drawing order.  If you add some
text in `on_idle()`, in an area that is not written, then it'll
be fine.

This is intended to be used when a `_view()` function, in lua,
doesn't draw any text of its own, so it should be OK for that purpose.

Closes #186.