Added a simple class to draw arbitrary modes.
This is the middle-layer from which our existing CFooView
classes will be updated to derive.

While it will always be possible to define new modes in
C++, and handle the display there, the existing modes are
so similar that having this middle-class makes sense to
cut down on duplication, at a small cost in complexity.

This updates #185