Moved status-panel code into a class of its own.
This removes the horrid `struct` we previously used in the
CScreen implementation, and because the status-panel is
now a class of its own we can remove the indirection which
was required to allow the Lua panel-object to access
the status-panel via that global CScreen object.

As a bonus we've fixed the drawing of the status-panel to
respond *natively* to resize events - without using a signal
handler.

This closes #200.

This closes #180.