When in simple mode draw empty lines, if appropriate.
This covers the case of a long line line which wraps beyond
the limit of the area to be drawn.

e.g:

     Subject: THis is a long line .. really long
     Date: March 10th 1976

     .. This is the end of the subject, which is still wrapped

The subject wraps past the end of the headers, because the message
has no body - which means the screen is cleared, the subject
is drawn, then because the body is empty it is left in-place
when it should be overwritten with a number of blank lines.