Merge two related functions.
We had defined two functions in lua:

        get_competiones
        function2table

Each of those called a function defined in lua that returned at
table - and presented that table as a std::vector of strings.

The only difference between them was that one function required
an argument (the token to complete) and one didn't.  Merge the
two routines on the observation that an empty argument is the
same as a missing argument.

Added new test-case to ensure things work as expected.