Allow multi-character keybindings.
This changes the input loop a little, rather than looking up
a C-string we use a std::string, and we build up that string
by appending the currently input-character if we're in the
middle of a multi-character input.

This code assumes that every key will either:

* Start a multi-key input.
* Complete a multi-key input.

This is not 100% perfect because it allows this:

    keymap['global']['^A^B'] = 'Panel:append( "This works." )'

But not this:

    keymap['global']['^A^B^C'] = 'Panel:append( "This works." )'

The assumption fails because "^A" would start a multi-charcter
input-value, but ^B would not complete it - so the input would
get trashed.

I think limiting ourselves to two characters is fine for the moment,
but we could revisit in the future if it is a problem.

This closes #212.