Removed get_nested_table.
I'm suspicious of table-related operations at the moment, as
a result of #210, so removing this horrid function is part of the
work of tracking down where we're failing.

Remove `get_nested_table`.  Replace that with a call to `lookup_key`,
which is implemented in Lua, for expanding bindings.

This closes #224.