Added debug information.
I'm trying to track down a leaking stack, and to do that I've
added tracing to show the current size of the stack as all of
our Lua functions are entered/exited.

At the moment I see a stack-overflow in the `call_sort` routine
which I think is because we're missing a pop at the end.  We'll
know for sure once I've added more logging though.

This is part of #225.