Don't update messages more often than we need to.
We find that opening a folder involves two transitions:

* Making the target folder the active one.
* Changing the mode to `index`.

Both of those operations will trigger a load of the messages, and
rightly-so.  But doing them in short space of time is pointless.

To avoid this we cache the contents based on the m_time of the
folder.  If the folder hasn't change,d and the mtime hasn't changed
then we can terminate early.

This closes #231.