Don't update the progress bar "too often".
Drawing the progress-bar unnecessarily is the cause of a slowdown,
which is ironic.  Since we know we're update the progress bar for
every column of the screen we can limit ourselves to only drawing
ourselves exactly that many times.

We just need to calculate the number of possible calls, ie the
things we're iterating over, and use:

        times = max % width

We use `math.fmod` to implement %, as per standard lua.

This updates #243.