Port to the C++ cache API.
THis updates our caching to use the C++-defined cache object,
rather than the pure Lua functionality, and is part of #249.

I suspect this means we're done, but I'll leave the bug open
for the moment:

* Cache is loaded on startup.
   *  If `cache.prefix` is defined.
* Cache is saved on exit.
   * `cache.prefix` will be created, if missing.

The cache will grow evermore, quoth the raven, but we can handle
that in the future by adding a timestamp to the cache-entries and
using LRU, or similar.