Allow format-string to format messages in index-view.
We now use the updated interpolation code we supplied
in the previous commit to allow the user to format the
index-view.

The default remains what it was:

        [${4|flags}] ${2|message_flags} - ${20|sender} - ${indent}${subject}

But the user may set `index.format` to use a different format.  The
most obvious change would be to use the senders name/email.  For
example:

        [${4|flags}] ${2|message_flags} - ${20|name} - ${indent}${subject}
        [${4|flags}] ${2|message_flags} - ${20|email} - ${indent}${subject}

The first shows `Steve Kemp`, the second shows `steve@steve.fi`, whereas
the default `sender` would show `Steve Kemp <steve@steve.fi>`.

This updates #268.