Avoid unreferening the message twice.
If we have a problem parsing the message the first
thing we do is this:

        /*
         * Clear our current object.
         */
        g_object_unref(stream);
        g_object_unref(parser);

But that is unnecessary because we've just done that immediately
above - the net result is we free the references *twice* which
causes problems.

Credit to Rudolf Bahr for pointing me in the direction of
this problem.