Allow message-saving to work.
If this function is called with an empty `dest` it prompts for one,
but it does so using a local-variable, which goes out of scope.

This means that ultimately Maildir-construction occurs with a null
destination path, and thus fails to save.

This updates #288.

I'll leave it open as I want to improve the function at the same
time.