Remove more (minor) leaks.
Specifically we were leaking message-headers, via the call to
g_mime_utils_header_decode_text, and we left a couple of
messages around.

This now allows me to open an index, read a few messages, and quit
with only a leak of ~200 bytes.  (Still need to work on that).

This updates #287