In the keybinding-list filter out "jump".
`jump(N)` is used to jump to line N of the display.  For example
if you open a maildir you might enter `300` to jump to line
300 - or the 300th email.

This is done by binding to the (multi-part) keypress 300.

Filter these bindings out of `keybinding_view()` to cut down on
the noise and make bindings more discoverable.

This was mentioned in #289.