Avoid drawing to an invisible window :)
This took a bit of tracking down, but there are times when
lumial2 will appear to hang, and this is around the time that the
status-panel is hidden.

Reproducer would be something like:

* Launch lumail2
* Toggle the size of the panel, via repeated pressing of `P`.
* Toggle the display of the panel, via repeated pressing of `TAB`.
* Press `a`, `t`, or `n` to limit the view of maildirs.

At some point in step 3-4 the program would hang, and it would
hang here:

     int CScreen::draw_single_line(int row, int col_offset, std::string
buf, WINDOW * screen, bool enable_scroll, bool enable_wrap)
     {

        ...
        ...

        getyx(screen, y, x);

        while (y == row)
        {
            /*
             * Draw " " over and over again, until we wrap and the row
             * changes to the next one.
             */
            waddstr(screen, (char *)" ");
            getyx(screen, y, x);

            count += 1;
        }

     }

The code here is saying "Get the current X/Y coords", "DRaw " "",
continue until we've bumped to the next row.

This code would constantly loop and the reason for that is that
the drawing would __not__ bump the cursor.  x,y would never change
because the window being drawn to was invisible and not present.

We've solved this problem by ensuring we skip the draw when the
window is invisible - triggered by a resize event - but we could
also add testing in the draw loop.

This updates #285.