Always install the new/current config-file.
The root cause of #293 was running with an old configuration-file
which had a bug in it.  Since the bug was fixed all should have
been well, but `make install` didn't forcibly install the
config-file.

(Because in the past we expected users to edit the main config
file, and blindly over-writing it would cause data-loss).

Nowadays we suggest the user leaves the main config-file
alone, and uses the `lumail2.user.lua` file as a template
for configuration.

To be more careful we'll archive the old config-file away,
based on the date, and then deploy the new one.

This closes #293 - even though it is unrelated.  :)