Changed install location of Lua lib + perl proxy.
This changes the installed location of things:

* Lua libraries are now at /usr/lib/lumail.
* Perl IMAP proxy is now at /usr/share/lumail.

The code is changed, the Makefile is changed, and the documentation
is changed.

This shouldn't break upgrades, but a warning is printed on startup
if the old locations are still present.  This will point users to:

        https://lumail.org/install/#upgrade

This closes #312.

It also updates #309, by sneakily making our new location refer
to `lumail` not `lumail2`, but note that no changes have been
made to the configuration-file(s), or the binary name in this
update.  We're solely changing the installation location(s).