Ensure that cache-keys don't contain spaces.
This commit changes the way that cache keys are used, such that they
are now silently updated to remove spaces.  This is required solely
because the "parser" that we use for (re)loading the cache from a
saved file tokenizes on spaces - throwing away values that contain
embedded spaces.

For example:

   $timestamp $key $value

If $key is set to "moi kiss" then instead of reading the value
correctly we store the value in the cache of:

        cache["moi"] = "kissa $value"

This is clearly wrong, and unfortunately isn't easily fixed.  We
could escape spaces but that would be complex.

This closes #341.