implement a simple header cache
The header cached is globaly active because it overwrites Message.header
to use the new cache.

The header caches for each maildir are stored in "cache.prefix"/hcache/path/to/maildir.

The header cache is saved when leaving a maildir and loaded when a new
one is entered.

The design of the header cache is described at the top of lib/header_cache.lua.