send utf-8 encoding header
Other email clients (only tested with thunderbird 52.8.0) have problems
with some utf8 runes (only tested with german umlauts) if this context
header is not added to the message prototype.

A encoding header is only missing if we don't use gmime to add another
mime part. If we add another mime part gmime maybe ignores our header or
is smart enough to use it. I can't see a difference in the messages with
an attachment before and after this change.