initialize m_running in screen.h
Before this commit m_running was initialized in CScreen:run_main_loop.
Files loaded with --load-file are executed before run_main_loop is
called. That means that if a file calls os.exit, which just sets
m_running to false, ha no effect because m_running is overwritten
in run_main_loop. Now it is possible to have a file do something with the
TUI and then exit immediately when its finished.